+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/subscription/SubscriptionHealthStatus.java
@@ -11,7 +11,7 @@ import io.vrap.rmf.base.client.JsonEnum;
 import io.vrap.rmf.base.client.utils.Generated;
 
 /**
- *  <p>The health status of the Subscription that indicates whether messages are being delivered to the Destination.</p>
+ *  <p>The health status of the Subscription that indicates whether messages are being delivered.</p>
  */
 @Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
 public interface SubscriptionHealthStatus extends JsonEnum {
@@ -28,6 +28,9 @@ public interface SubscriptionHealthStatus extends JsonEnum {
     /**
     <p>Does not deliver messages temporarily due to reasons other than a configuration error. For example, the Destination has a temporary outage.</p> */
     SubscriptionHealthStatus TEMPORARY_ERROR = SubscriptionHealthStatusEnum.TEMPORARY_ERROR;
+    /**
+    <p>Does not deliver messages with the current configuration and the delivery of the messages is no longer attempted. Undelivered messages are not retained and will not be delivered. The <code>status</code> will not automatically change to Healthy. To return your subscriptions to a Healthy status, please contact our support team.</p> */
+    SubscriptionHealthStatus MANUALLY_SUSPENDED = SubscriptionHealthStatusEnum.MANUALLY_SUSPENDED;
 
     /**
      * possible values of SubscriptionHealthStatus
@@ -51,7 +54,12 @@ public interface SubscriptionHealthStatus extends JsonEnum {
         /**
          * TemporaryError
          */
-        TEMPORARY_ERROR("TemporaryError");
+        TEMPORARY_ERROR("TemporaryError"),
+
+        /**
+         * ManuallySuspended
+         */
+        MANUALLY_SUSPENDED("ManuallySuspended");
         private final String jsonName;
 
         private SubscriptionHealthStatusEnum(final String jsonName) {